requirements.txt
livekit-agents[google,sarvam]
livekit-plugins-silero
python-dotenv
